[FIX] Fix tests and actions

+++ e-budget-be/src/main/java/com/ebudget/account/resource/request/NewAccountDTO.java
@@ -4,6 +4,8 @@ import com.ebudget.account.model.enumeration.AccountType;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record NewAccountDTO(
         String accountLogo,
         @NotBlank
@@ -11,6 +13,6 @@ public record NewAccountDTO(
         @NotNull
         AccountType accountType,
         @NotNull
-        Double initialBalance
+        BigDecimal initialBalance
 ) {
 }
